Fix < 1ms TimeSpan's to be Sleep(0)
Instead of calling out to the Millisecond based sleep method, it's
better to call to the Microsecond based sleep method.

```
var ts = TimeSpan.FromMilliseconds(.5);
var millis = Convert.ToInt64(ts.TotalMilliseconds);
var micros = ts.Ticks / (TimeSpan.TicksPerMillisecond / 1000);
```

In this example, `millis == 0`, and `micros == 500`.

If `Sleep(TimeSpan` was called with that `ts` it would effectively be
`Sleep(0)` which can be problematic.